Fix the `TestForkServer::testExecuteProcess' test case under 64 bits systems.
An u_long is not 4 bytes long under 64 bits systems, use int instead.

+++ myserver/src/base/process/fork_server.cpp
@@ -97,7 +97,7 @@ int ForkServer::readInt (Socket *sock, int *dest)
  */
 int ForkServer::readString (Socket *sock, char **out)
 {
-  u_long len;
+  int len;
   u_long nbr;
 
   if (sock->read ((char*) &len, 4, &nbr) || nbr < 4)
